Keep desktop integrations and executable Forge plugins out of iOS
