+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to Moldavite are documented here.
 
 ### Fixed
 
+- iOS no longer exposes desktop software updates, plugin execution or installation, calendar connections, WordPress publishing, or PDF export.
+
 - iOS onboarding uses the device’s local Forge, introduces touch navigation, and stays visible above open pages. Desktop-only folder selection and AI feature pages stay out of the phone flow.
 
 ## [2.6.0] - 2026-09-04
