+++ docs/PROJECT_STATUS.md
@@ -52,7 +52,7 @@
 - macOS builds are signed and notarized. Windows installers are unsigned and may trigger SmartScreen because they are not Authenticode-signed. Linux bundles are unsigned as well; Linux has no equivalent warning. Updater artifacts are signed for every platform, including Windows, and clients verify them before installation. Checks run about 15 seconds after launch, every 24 hours while open, and on focus after 24 hours without a successful check, and can be switched off in Settings → About (manual checks still work); automatic network/404 failures stay silent and retry, while pending versions add accent dots to Settings and About plus the existing install action. Manual checks retain explicit errors, and completed upgrades show the CHANGELOG-backed "What's New" popup (see docs/RELEASING.md)
 - Themes/presets, platform-specific keyboard shortcut labels and overlay (⌘? on macOS, Ctrl+? on Windows and Linux), settings modal with focus trap
 - Window size and position are restored between launches
-- iOS is in development on the `mobile/ios-spike` branch, not shipped: the same React app and Rust core build through Tauri's iOS target and run on the iPhone simulator with a phone layout (icon rail navigation, full-screen pages, two-level Settings, keyboard-aware shell, local-container onboarding), a home screen widget that opens today's note through `moldavite://today`, and the desktop-only parts compiled out. See docs/MOBILE.md for the build, the gates and what is left
+- iOS is in development on the `mobile/ios-spike` branch, not shipped: the same React app and Rust core build through Tauri's iOS target and run on the iPhone simulator with a phone layout (icon rail navigation, full-screen pages, two-level Settings, keyboard-aware shell, local-container onboarding), a home screen widget that opens today's note through `moldavite://today`, and the desktop-only parts compiled out. Third-party plugins and their installation commands are desktop-only. See docs/MOBILE.md for the build, the gates and what is left
 
 ### Browser clipper
