+++ docs/plugins.html
@@ -924,7 +924,8 @@ api.commands.<span class="f">add</span>({
           <span class="title">Open the bundled examples on GitHub →</span>
         </a>
       </div>
-    </main>
+    <p>Plugins are desktop-only. The iOS app in development does not install or run Forge plugins.</p>
+</main>
 
     <footer class="site-footer">
       <div class="footer-inner">
