+++ src-tauri/src/commands/mod.rs
@@ -18,6 +18,7 @@ pub mod locking;
 pub mod mcp_settings;
 pub mod misc;
 pub mod notes;
+#[cfg(desktop)]
 pub mod plugins;
 pub mod root_files;
 pub mod search;
