+++ src/components/editor/MoreOptionsMenu.tsx
@@ -1,3 +1,4 @@
+import { isMobilePlatform } from '@/lib/platform';
 import { lazy, Suspense, useState } from 'react';
 import { useShallow } from 'zustand/react/shallow';
 import { save } from '@tauri-apps/plugin-dialog';
@@ -271,7 +272,9 @@ export function MoreOptionsMenu({
           <DropdownItem onClick={() => setShowRenameModal(true)}>Rename note…</DropdownItem>
         )}
         <DropdownItem onClick={handleExport}>Export as Markdown</DropdownItem>
-        <DropdownItem onClick={handleExportPdf}>Export as PDF…</DropdownItem>
+        {!isMobilePlatform() && (
+          <DropdownItem onClick={handleExportPdf}>Export as PDF…</DropdownItem>
+        )}
         <DropdownItem onClick={handleExportPlaintext}>Export as Plaintext</DropdownItem>
         <DropdownItem onClick={() => setShowSaveTemplateModal(true)}>Save as template</DropdownItem>
         <DropdownDivider />
